pyproject.toml
[project]
name = "casl"
version = "0.1.0"
description = "CASL Course Document Generator - Extract data from Excel and generate Word/Markdown documents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "fpdf2>=2.8.0",
    "openpyxl>=3.1.0",
    "pydantic>=2.0.0",
    "python-docx>=1.1.0",
    "streamlit>=1.40.0",
]
